Fix screenshot pixmap being refreshed on each paint event, fixes #106
On X11 systems QPixmap's data is implicitly shared so that any external
change to it is propagated each time the pixmap is accessed. This is exactly
the case when SelectionOverlay widget's paint event handler is writing
it to the widget's painter resulting in effects as shown in #106.

In order to fix it the pixmap can be deep-copied at the moment of
grabbing the screen so that it's not connected to the original X11
pixmap anymore.